Prepare JSONThunker for Py3 and Move JSONThunker tests
